finished migreating icons to common-elements/icons, moved a few one off svgs to the files they are used in, will want to refactor later to avoid 1 off icons
