+++ packages/augur-ui/src/modules/alerts/components/alerts-view.styles.less
@@ -70,6 +70,10 @@
   top: @size-16;
   width: @size-16;
 
+  > svg > g > g {
+    stroke: #F6F6F8;
+  }
+
   @media @breakpoint-mobile {
     display: block;
   }
