+++ packages/augur-ui/src/modules/auth/components/common/error-container.tsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import classNames from "classnames";
 
 import { ACCOUNT_TYPES, ERROR_TYPES } from "modules/common-elements/constants";
-import { errorIcon } from "modules/common/components/icons";
+import { errorIcon } from "modules/common-elements/icons";
 
 import StylesDropdown from "modules/auth/components/connect-dropdown/connect-dropdown.styles";
 import Styles from "modules/auth/components/common/error-container.styles";
