+++ packages/augur-ui/src/modules/auth/components/common/hardware-wallet.tsx
@@ -10,7 +10,7 @@ import classNames from "classnames";
 import AddressPickerContent from "modules/auth/components/common/address-picker-content";
 import DerivationPathEditor from "modules/auth/components/common/derivation-path-editor";
 import { ERROR_TYPES } from "modules/common-elements/constants";
-import { errorIcon } from "modules/common/components/icons";
+import { errorIcon } from "modules/common-elements/icons";
 import { filter } from "lodash";
 import Styles from "modules/auth/components/common/hardware-wallet.styles";
 import StylesDropdown from "modules/auth/components/connect-dropdown/connect-dropdown.styles";
