+++ packages/augur-ui/src/modules/common/less/tooltip.styles.less
@@ -39,6 +39,15 @@ div.Tooltip {
   > svg {
     max-height: @size-16;
     max-width: @size-16;
+
+    > circle { 
+      fill: @color-tertiary-one;
+      stroke: @color-primary-alternate;
+    }
+
+    > path {
+      fill: @color-secondary-text;
+    }
   }
 }
