+++ packages/augur-ui/src/modules/create-market/components/create-market-form-define/create-market-form-define.tsx
@@ -10,7 +10,7 @@ import {
   TAGS_MAX_LENGTH
 } from "modules/common-elements/constants";
 
-import { ExclamationCircle as InputErrorIcon } from "modules/common/components/icons";
+import { InputErrorIcon } from "modules/common-elements/icons";
 import moment from "moment";
 import Styles from "modules/create-market/components/create-market-form-define/create-market-form-define.styles";
 import StylesForm from "modules/create-market/components/create-market-form/create-market-form.styles";
