+++ packages/augur-ui/src/modules/create-market/components/create-market-form-liquidity/create-market-form-liquidity.tsx
@@ -8,7 +8,7 @@ import { BigNumber, createBigNumber } from "utils/create-big-number";
 import { augur } from "services/augurjs";
 
 import InputDropdown from "modules/common/components/input-dropdown/input-dropdown";
-import { ExclamationCircle as InputErrorIcon } from "modules/common/components/icons";
+import { InputErrorIcon } from "modules/common-elements/icons";
 import CreateMarketFormLiquidityCharts from "modules/market-charts/containers/create-market-form-liquidity-charts";
 
 import {
