+++ packages/augur-ui/src/modules/create-market/components/create-market-form-outcome/create-market-form-outcome.tsx
@@ -19,9 +19,9 @@ import {
 } from "modules/common-elements/constants";
 
 import {
-  ExclamationCircle as InputErrorIcon,
+  InputErrorIcon,
   Hint
-} from "modules/common/components/icons";
+} from "modules/common-elements/icons";
 import Styles from "modules/create-market/components/create-market-form-outcome/create-market-form-outcome.styles";
 import StylesForm from "modules/create-market/components/create-market-form/create-market-form.styles";
 import ReactTooltip from "react-tooltip";
