+++ packages/augur-ui/src/modules/create-market/components/create-market-form-resolution/create-market-form-resolution.tsx
@@ -12,7 +12,7 @@ import {
 } from "modules/common-elements/constants";
 import isAddress from "modules/auth/helpers/is-address";
 
-import { ExclamationCircle as InputErrorIcon } from "modules/common/components/icons";
+import { InputErrorIcon } from "modules/common-elements/icons";
 import Styles from "modules/create-market/components/create-market-form-resolution/create-market-form-resolution.styles";
 import StylesForm from "modules/create-market/components/create-market-form/create-market-form.styles";
