+++ packages/augur-ui/src/modules/create-market/components/create-market-form/create-market-form.tsx
@@ -12,7 +12,7 @@ import CreateMarketLiquidity from "modules/create-market/components/create-marke
 import CreateMarketLiquidityOrders from "modules/create-market/components/create-market-form-liquidity-orders/create-market-form-liquidity-orders";
 import CreateMarketReview from "modules/create-market/components/create-market-form-review/create-market-form-review";
 import Styles from "modules/create-market/components/create-market-form/create-market-form.styles";
-import { ExclamationCircle as InputErrorIcon } from "modules/common/components/icons";
+import { InputErrorIcon } from "modules/common-elements/icons";
 import { createBigNumber } from "utils/create-big-number";
 import { CATEGORICAL, SCALAR, BID } from "modules/common-elements/constants";
 import moment from "moment";
