+++ packages/augur-ui/src/modules/filter-sort/components/filter-search/filter-search.tsx
@@ -7,7 +7,7 @@ import makeQuery from "modules/routes/helpers/make-query";
 
 import { PAGINATION_PARAM_NAME } from "modules/routes/constants/param-names";
 import { FILTER_SEARCH_PARAM } from "modules/common-elements/constants";
-import { Hint } from "modules/common/components/icons";
+import { Hint } from "modules/common-elements/icons";
 import Styles from "modules/filter-sort/components/filter-search/filter-search.styles";
 import ReactTooltip from "react-tooltip";
 import TooltipStyles from "modules/common/less/tooltip.styles";
