+++ packages/augur-ui/src/modules/market/components/market-header/market-header.tsx
@@ -3,10 +3,10 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import classNames from "classnames";
 import {
+  BackArrow,
   ChevronDown,
   ChevronUp
-} from "modules/common/components/icons";
-import { BackArrow } from "modules/common-elements/icons";
+} from "modules/common-elements/icons";
 import MarkdownRenderer from "modules/common-elements/markdown-renderer";
 import MarketHeaderBar from "modules/market/containers/market-header-bar";
 import { BigNumber } from "bignumber.js";
