+++ packages/augur-ui/src/modules/modal/components/modal-participate.tsx
@@ -3,7 +3,7 @@
 import React, { Component } from "react";
 import classNames from "classnames";
 import { createBigNumber } from "utils/create-big-number";
-import { ExclamationCircle as InputErrorIcon } from "modules/common/components/icons";
+import { InputErrorIcon } from "modules/common-elements/icons";
 import Input from "modules/common/components/input/input";
 import ModalActions from "modules/modal/components/common/modal-actions";
 import ModalReview from "modules/modal/components/modal-review";
