+++ packages/augur-ui/src/modules/reporting/components/reporting-report-form/reporting-report-form.tsx
@@ -9,7 +9,7 @@ import { constants } from "services/augurjs";
 import { YES_NO, CATEGORICAL, SCALAR } from "modules/common-elements/constants";
 import FormStyles from "modules/common/less/form";
 import Styles from "modules/reporting/components/reporting-report-form/reporting-report-form.styles";
-import { ExclamationCircle as InputErrorIcon } from "modules/common/components/icons";
+import { InputErrorIcon } from "modules/common-elements/icons";
 import { formatRep } from "utils/format-number";
 import { RepBalance } from "modules/common-elements/labels";
