+++ packages/augur-ui/src/utils/toggle-height.styles.less.d.ts
@@ -1,0 +1,4 @@
+export const target: string;
+export const open: string;
+export const quick: string;
+export const openOnMobile: string;
